requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
Werkzeug==2.3.7
python -m venv attendance_env

